Add custom hook and update components to use it
