+++ src/hooks/useLocalStorageState.js
@@ -1,0 +1,14 @@
+import { useState, useEffect } from 'react';
+
+export const useLocalStorageState = (initialState, key) => {
+  const [value, setValue] = useState(() => {
+    const saved = localStorage.getItem(key);
+    return saved ? JSON.parse(saved) : initialState;
+  });
+
+  useEffect(() => {
+    localStorage.setItem(key, JSON.stringify(value));
+  }, [value, key]);
+
+  return [value, setValue];
+};
